+++ src/features/codeExplainer/ui/index.html
@@ -9,6 +9,7 @@
 </head>
 <body>
   <div class="container">
+    
     <!-- HEADER -->
     <div class="header">
       <div class="title-group">
